+++ src/Main.java
@@ -4,8 +4,8 @@ public class Main {
     public static void main(String[] args) {
         BmiService service = new BmiService();
         double height = 1.87; // метры
-        double weight = 98; // кг
-        double Bmi = service.calculate(height, weight); //должно получиться 28
+        int weight = 98; // кг
+        int Bmi = service.calculate(height, weight); //должно получиться 28
         System.out.printf("индекс массы тела " + Bmi);
